% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_model_to_db}
\alias{add_model_to_db}
\title{Add nomad model to package model_db}
\usage{
add_model_to_db(nomad_model, name = deparse(substitute(nomad_model)))
}
\arguments{
\item{nomad_model}{A \code{\link[=nomad_model]{nomad_model()}} object}

\item{name}{Name of the nomad model to be added. If blank, the name of the
\code{nomad_model} parameter will be uses. The name should abide to the
naming conventions detailed in the Data vignette, i.e. it should look
similar to zmb_cdr_2020_mod_dd_exp in structure}
}
\description{
\code{add_model_to_db} correctly adds new nomad models to the internal
\link{model_db} database. You should only need to use this function
if you are planning on adding a new model to the \code{nomad} package and you
have read the \url{https://ojwatson.github.io/nomad/articles/data.html}
vignette.
}
