\name{Balf2}
\docType{data}
\alias{Balf2}
\title{Second Run of Balf Experiment}
\description{
  MS peptide quantiation for second run of Balf experiment
}
\usage{data(BalfPeptides)}
\value{data.frame of MS peptide quantification of second run of Balf 
experiment}
\format{Text file with space separator}
\references{
  Bhargava,M. et al. (2014) Proteomic profiles in acute respiratory distress syndrome differentiates survivors from non-survivors. PloS One, 9, e109713. 
}
