\name{Balf3}
\docType{data}
\alias{Balf3}
\title{Third Run of Balf Experiment}
\description{
  MS peptide quantiation for third run of Balf experiment
}
\value{data.frame of MS peptide quantification of third run of Balf experiment}
\usage{data(BalfPeptides)}
\format{Text file with space separator}
\references{
  Bhargava,M. et al. (2014) Proteomic profiles in acute respiratory distress syndrome differentiates survivors from non-survivors. PloS One, 9, e109713. 
}
