\name{BalfPeptides}
\docType{data}
\alias{BalfPeptides}
\title{Balf Experiment Data}
\description{
BalfPeptides is a reformat of Balf1 to Balf3 to a from friendly to NOMAD.
}
\value{
Data.frame of MS peptide quantification.
}
\usage{data(BalfPeptides)}
\format{R data objects}
\references{
  Bhargava,M. et al. (2014) Proteomic profiles in acute respiratory distress syndrome differentiates survivors from non-survivors. PloS One, 9, e109713. 
}
