% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{check}
\alias{check}
\title{Check goodness of fit for 'nomad_model' class}
\usage{
check(object, plots, ...)
}
\arguments{
\item{object}{a \code{\link[=nomad_model]{nomad_model()}} object containing the fitted mobility}

\item{plots}{logical indicating whether to plot the Posterior Predictive Check and Normal Q-Q Plot (default = \code{TRUE})}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list of goodness of fit measures
}
\description{
This function takes a \code{nomad_model} object and calculates goodness of fit
metrics for the underlying \code{\link[mobility:mobility]{mobility::mobility()}}. If the Deviance
Information Criterin (DIC) was calculated in the supplied model object,
it is included in output. When plots = TRUE, two plots are shown containing
the posterior distribution of trip counts compared to observed data and a
Normal Q-Q plot showing the quantiles of model residuals against those
expected from a Normal distribution.
}
\details{
Goodness of fit metrics include:
\describe{
  \item{DIC}{\href{https://en.wikipedia.org/wiki/Deviance_information_criterion}{Deviance Information Criterion}}
  \item{RMSE}{\href{https://en.wikipedia.org/wiki/Root-mean-square_deviation}{Root Mean Squared Error}}
  \item{MAPE}{\href{https://en.wikipedia.org/wiki/Mean_absolute_percentage_error}{Mean Absolute Percent Error}}
  \item{R2}{\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{R-squared}}
}
}
\examples{
# Get nomad_model object
nmd_model <- nomad::model_db$zmb_cdr_2020_mod_dd_exp

# Check model fit
nomad::check(nmd_model)

# Get nomad_model object without underlying data
nmd_model <- nomad::model_db$zmb_fb_2020_mod_grav_exp

# Model check statistics are still available as these are
# saved when model data is removed from object
nomad::check(nmd_model)
}
\seealso{
Other model: 
\code{\link[mobility]{compare}()},
\code{\link[mobility]{fit_jags}()},
\code{\link[mobility]{fit_prob_travel}()},
\code{\link[mobility]{mobility}()},
\code{\link[mobility]{predict}()},
\code{\link[mobility]{residuals}()},
\code{\link[mobility]{summary}()}
}
\author{
John Giles
}
