% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{mobility_metrics}
\alias{mobility_metrics}
\title{Summarise predicted mobility matrices}
\usage{
mobility_metrics(x)
}
\arguments{
\item{x}{A predicted mobility matrix, or a 3D array with simulations in the
third dimension.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per matrix.
}
\description{
Calculates simple summary metrics for one predicted matrix or each simulated
matrix in a 3D prediction array.
}
