% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_db}
\alias{model_db}
\title{Model database}
\format{
A named \link{list} of \link{nomad_model}s.

Each model is named with a standard convention. For example, let's look at
\code{zmb_cdr_2020_mod_dd_exp}. Each of the names in the snake case name convention
refer either to the data that the model was fit using or the type of model
that was fit.

\code{zmb_cdr_2020_mod_dd_exp}:
\itemize{
\item \code{zmb}: ISO3C county code for where data was collected.
\item \code{cdr}: type of mobility data. cdr = Call Data Record.
\item \code{2020}: year of data collection.
\item \code{mod}: signifies the following labels relate to the model.
\item \code{dd}: mobility model. dd = Departure-Diffusion.
\item \code{exp}: sub-type of mobility model. exp = Exponential.
}

The naming conventions help with documenting the models and enable linking
to \link{mobility_db} to query further information about the underlying
mobility data the model was fit using. All names in the snake case model name
before \code{mod} refer to the mobility data.
}
\usage{
model_db
}
\description{
Database of all mobility models stored in \code{nomad}
}
\keyword{datasets}
