\name{NOMAD-package}
\docType{package}
\title{
ANOVA normalization of MS iTRAQ peptide abundances
}
\description{
This package contains functions for performing an ANOVA normalization
method to MS iTRAQ peptide abundances and assembling protein abundances.
}
\details{
\tabular{ll}{
Package: \tab NOMAD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-15\cr
License: \tab GPL (>=2) \cr
}
nomadNormalization ApplyANOVA normalization to peptide abundances
nomadAssembleProteins Assemble protein abundances from peptide
abundances
nomadCheckBias Visualize bias of a particular factor
nomadCheckLogTransform Visualize effect of logging on homogeneity of variance
}
\author{
Carl Murie

Maintainer: Carl Murie <carl.murie@ki.se>
}
\references{
Hill,E.G. et al. (2008) A Statistical Model for iTRAQ Data Analysis. J. Proteome Res., 7, 3091-3101.
Oberg,A.L. et al. (2008) Statistical Analysis of Relative Labeled Mass Spectrometry Data from Complex Samples Using ANOVA. J. Proteome Res., 7, 225-233.
}

\keyword{ package }
