% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{nomad_model}
\alias{nomad_model}
\title{Create nomad_model}
\usage{
nomad_model(model, data_name, data_ref = NULL, data_fields = NULL)
}
\arguments{
\item{model}{\code{\link[mobility:mobility]{mobility::mobility()}} model}

\item{data_name}{The model's name to link to \code{nomad::mobility_db}}

\item{data_ref}{Optional internal reference to shared package model data.}

\item{data_fields}{Optional character vector of shared data fields to attach
when the model is hydrated.}
}
\value{
A new \code{nomad_model} object.
}
\description{
A nomad mobility model
}
\examples{
# Get mobility.model object
mob_model <- nomad::model_db$zmb_fb_2020_mod_grav_exp$get_model()

# Create nomad_model object
new_mod <- nomad_model(mob_model, data_name = "zmb_fb_2020")
}
