% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOMAD.R
\name{nomadAssembleProteins}
\alias{nomadAssembleProteins}
\title{Assemble Protein abundance measurements from Peptide abundance
measurements}
\usage{
nomadAssembleProteins(y, x, method = "TukeyBW", format = "BySample",
  combineDupPeptides = FALSE, standardizeAbundance = FALSE,
  mySep = "&_MYSEP_&")
}
\arguments{
\item{y}{\code{vector} Dependant variable (vector of real numbers)}

\item{x}{\code{vector} Independent or explanatory variables (matrix). In this context they
will be Peptide, Protein, Run, and iTRAQ identifiers.
Run and iTRAQ identifiers must be numeric}

\item{method}{\code{string} Method of calculating center of peptide abundances per protein
- choices: TukeyBW, Median, or Mean. Default is TukeyBW}

\item{format}{\code{string} BySample - each column is an individual sample
ByProtein - rows are proteins-columns are abundance,day and itraq}

\item{combineDupPeptides}{\code{string} If TRUE then average (based on method) duplicated
peptides for the same protein before calculating protein
abundance}

\item{standardizeAbundance}{\code{logical} If TRUE then transform each sample (experiment and
iTRAQ) with a robust Z adjustment. The default is
FALSE.}

\item{mySep}{\code{logical} Character string that must not exist in any factor in "x".}
}
\value{
\code{List} with elements \code{scores} matrix of normalized protein scores. 
each row is a protein and each column is a single sample (run and iTRAQ)
}
\description{
Assemble Protein abundance measurements from Peptide abundance
measurements
}
\examples{
data(BalfPeptides)

ret <- nomadNormalization(y=BalfPeptides$Abundance, x=BalfPeptides)
scores <- nomadAssembleProteins(ret$y, ret$x)

}
