% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOMAD.R
\name{nomadCheckBias}
\alias{nomadCheckBias}
\title{Check level of bias based on batch effect.}
\usage{
nomadCheckBias(dat, fact, numNAs = dim(dat)[[2]], label = NULL,
  yMax = NULL, showSampleSize = TRUE)
}
\arguments{
\item{dat}{\code{matrix} Rows are protein, columns are samples (each iTRAQ sample 
per day. The same format as the "scores" element of the output of
assembleProteins with format=bySample. Missing data must be shown
as NA.}

\item{fact}{\code{vector} numeric or character vector of labels identifying batches.
Generally a batch is a single MS run of 4 or 8 iTRAQ channels.}

\item{numNAs}{\code{integer} Number of missing data points allowed. Defaults to number of
samples in data}

\item{label}{\code{string} Text for title of graph}

\item{yMax}{\code{real} Upper count range for plotting histograms.}

\item{showSampleSize}{\code{boolean} If TRUE add text box showing number of samples.}
}
\value{
None
}
\description{
Check level of bias based on batch effect.
}
\examples{
data(BalfPeptides)

## assemble the protein scores without normalization
scores <- nomadAssembleProteins(BalfPeptides$Abundance, BalfPeptides)
runInd <- rep(1:3, each=8)  ## each integer is a single batch (3 batches total)
nomadCheckBias(scores$scores, fact=runInd)
}
