% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOMAD.R
\name{nomadCheckLogTransform}
\alias{nomadCheckLogTransform}
\title{Check whether logging the data is necessary. y and x are the same inputs
as nomadNormalization
Check whether logging the data is necessary}
\usage{
nomadCheckLogTransform(y, x, meth = "TukeyBW", rawTrim = 1)
}
\arguments{
\item{y}{\code{vector} dependant variable (vector of real numbers)}

\item{x}{\code{matrix} independent or explanatory variables (matrix). x must include columns
for Peptide, Protein, iTRAQ, and Run}

\item{meth}{\code{string} manner of calculating center of peptide abundances per protein
- choices: TukeyBW, Median, or Mean}

\item{rawTrim}{\code{real} Quantile used to reduce scale for plotting raw data. Will plot
raw means up to the specified quantile (default=1)}
}
\value{
None
}
\description{
Check whether logging the data is necessary. y and x are the same inputs
as nomadNormalization
Check whether logging the data is necessary
}
\examples{
data(BalfPeptides)
nomadCheckLogTransform(y=BalfPeptides$Abundance, x=BalfPeptides)

}
