% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOMAD.R
\name{nomadNormalization}
\alias{nomadNormalization}
\title{Apply anova model to mass spec data. Use residuals of anova model as
final scores.}
\usage{
nomadNormalization(y, x, factors = NULL, doRobust = TRUE, doLog = TRUE,
  doiTRAQCorrection = FALSE, iTRAQCorrectionTable = NULL)
}
\arguments{
\item{y}{\code{real} Dependant variable (vector of positive real numbers)}

\item{x}{\code{matrix} Independent or explanatory variables (matrix). In this context they
will be "Peptide","Protein", "Run", and "iTRAQ" identifiers. An
optional "iTRAQCorrectionIndex" column is necessary if iTRAQ isotope
correction is desired.}

\item{factors}{\code{vector} Default is to use the 4 main factors (Peptide, Protein,
Run, and iTRAQ)  and the three interactions between
Run and the other three main effects. The user can
define their own factors. The "factors" list must match the
data.frame labels of x.}

\item{doRobust}{\code{logical} If TRUE use medians of levels instead of the default means.
Default is to use median.}

\item{doLog}{\code{logical} If TRUE then y will be logged (log 2). Default is to log.}

\item{doiTRAQCorrection}{\code{logical} If true then orrect peptide abundance scores based on
 their iTRAQ labelling. Default is TRUE. iTRAQ
correction values are hardcoded for 8 iTRAQs. The
iTRAQCorrectionTable must be provided for iTRAQs that
do not number 8.}

\item{iTRAQCorrectionTable}{\code{matrix} Matrix of iTRAQ corrections. If iTRAQ equals 8
then an internal correction table is used. If a
table is provided when iTRAQ equals 8 then the
user supplied table will be used. If the number
of iTRAQs is not eight than a table must be
supplied by the user.}
}
\value{
\code{List} elements are \code{y}: normalized scores after anova normalization,
\code{x}: design matrix, \code{removedData}: index of rows removed from input data,
\code{call}: function call.
}
\description{
Apply anova model to mass spec data. Use residuals of anova model as
final scores.
}
\examples{
data(BalfPeptides)

ret <- nomadNormalization(y=BalfPeptides$Abundance, x=BalfPeptides)

}
