% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{predict.nomad_ensemble}
\alias{predict.nomad_ensemble}
\title{Predict from a nomad model ensemble}
\usage{
\method{predict}{nomad_ensemble}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A \code{\link[=nomad_ensemble]{nomad_ensemble()}} object.}

\item{newdata}{Optional prediction data passed to each model.}

\item{...}{Further arguments passed to \code{\link[=predict.nomad_model]{predict.nomad_model()}}.}
}
\value{
A weighted-average prediction matrix or array.
}
\description{
Predict from a nomad model ensemble
}
