% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{prediction_difference}
\alias{prediction_difference}
\title{Compare two prediction matrices}
\usage{
prediction_difference(x, y, relative = FALSE)
}
\arguments{
\item{x, y}{Prediction matrices with the same dimensions.}

\item{relative}{If \code{TRUE}, return \code{(x - y) / y}; otherwise return \code{x - y}.}
}
\value{
A matrix of differences.
}
\description{
Compare two prediction matrices
}
