% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{print.nomad_model}
\alias{print.nomad_model}
\title{Print 'nomad_model' class}
\usage{
\method{print}{nomad_model}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=nomad_model]{nomad_model()}} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
This function provides a simple print overview of a \code{nomad_model} object.
}
