% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{print}
\alias{print}
\title{Print 'nomad_model' class}
\usage{
print(x, name, ...)
}
\arguments{
\item{x}{a \code{\link[=nomad_model]{nomad_model()}} object}

\item{name}{Name of the model. Defaults to the name of object x}

\item{...}{further arguments passed to or from other methods}
}
\description{
This function provides a simple print overview of a \code{nomad_model} object.
}
