% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print_mobility_data}
\alias{print_mobility_data}
\title{Clean print information on a mobility dataset}
\usage{
print_mobility_data(data_name)
}
\arguments{
\item{data_name}{The model's name to link to \link{mobility_db}}
}
\description{
Clean print information on a mobility dataset
}
