% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{R6_nomad_model}
\alias{R6_nomad_model}
\title{R6 Class for mobility models}
\description{
A nomad mobility model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-nomad.model-new}{\code{R6_nomad_model$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad.model-new"></a>}}
\if{latex}{\out{\hypertarget{method-nomad.model-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{nomad} mobility model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_nomad_model$new(model, model_desc, data_desc, model_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{\link[mobility:mobility]{mobility::mobility()}} model}

\item{\code{model_desc}}{Short description of model type and underlying data}

\item{\code{data_desc}}{Short description of data used to fit \code{model}}

\item{\code{model_name}}{The name of the model for linking with \code{nomad::model_data}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{nomad.model} object.
}
}
}
