% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals.nomad_model}
\alias{residuals.nomad_model}
\title{Extract model residuals 'nomad_model' class}
\usage{
\method{residuals}{nomad_model}(object, type = "deviance", ...)
}
\arguments{
\item{object}{a \code{\link[=nomad_model]{nomad_model()}} object containing the fitted mobility}

\item{type}{the type of residuals to be returned. Available residual types include: \code{'deviance'} (default), \code{'pearson'}, and \code{'raw'}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
a matrix containing model residuals
}
\description{
This function takes a \code{nomad_model} object and extracts model residuals from
the underlying \code{\link[mobility:mobility]{mobility::mobility()}}.
}
\details{
Residual types are calculated as:
\describe{
  \item{raw}{\eqn{y_i - \mu_i}}
  \item{pearson}{\eqn{(y_i - \mu_i)/\sqrt{\mu_i}}}
  \item{deviance}{\eqn{sign(y_i - \mu_i) * \sqrt(2(log(y_i/\mu_i) - (y_i - \mu_i)))}}
}
Where, \eqn{y_i} is the observed data and \eqn{\mu_i} is the value predicted by the model using the mean of parameter posterior distributions
}
\examples{
# Get nomad_model object
nmd_model <- nomad::model_db$zmb_cdr_2020_mod_dd_exp

# Get model residuals
residuals(nmd_model)

# Get nomad_model object without underlying data
nmd_model <- nomad::model_db$zmb_fb_2020_mod_grav_exp

# Model residuals not available
residuals(nmd_model)
}
\seealso{
Other model: 
\code{\link[mobility]{check}()},
\code{\link[mobility]{compare}()},
\code{\link[mobility]{fit_jags}()},
\code{\link[mobility]{fit_prob_travel}()},
\code{\link[mobility]{mobility}()},
\code{\link[mobility]{predict}()},
\code{\link[mobility]{summary}()}
}
\author{
John Giles
}
