% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOMAD.R
\name{solveiTRAQCorrection}
\alias{solveiTRAQCorrection}
\title{Correct peptide abundances based on iTRAQ specification. The iTRAQTable,
below characterizes the components of the abundance scores.}
\usage{
solveiTRAQCorrection(y, iTRAQTable = NULL)
}
\arguments{
\item{y}{\code{vector} A vector of n peptide abundances ordered by iTRAQ (n iTRAQ labels)}

\item{iTRAQTable}{\code{matrix} n by n matrix of isotope corrections}
}
\value{
vector of peptide abundances after correcting for isotope bleed over.
}
\description{
Correct peptide abundances based on iTRAQ specification. The iTRAQTable,
below characterizes the components of the abundance scores.
}
