% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOMAD.R
\name{solveiTRAQCorrection4}
\alias{solveiTRAQCorrection4}
\title{Correct peptide abundances based on iTRAQ specification. The iTRAQTable
below characterizes the components of the abundance scores.}
\usage{
solveiTRAQCorrection4(y, iTRAQTable = NULL)
}
\arguments{
\item{y}{\code{vector} A vector of 4 peptide abundances ordered by iTRAQ (1 to 4)}

\item{iTRAQTable}{\code{matrix} An iTRAQTable - 4x4 matrix of isotope corrections. Probably shouldn't modify
unless you have a very good reason.}
}
\value{
vector of peptide abundances after correcting for isotope bleed over.
}
\description{
Correct peptide abundances based on iTRAQ specification. The iTRAQTable
below characterizes the components of the abundance scores.
}
