% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{summary.nomad_model}
\alias{summary.nomad_model}
\title{Calculate summary statistics for 'nomad_model' class}
\usage{
\method{summary}{nomad_model}(object, probs = c(0.025, 0.975), ac_lags = c(5, 10), ...)
}
\arguments{
\item{object}{a \code{\link[=nomad_model]{nomad_model()}} object containing the fitted mobility}

\item{probs}{numeric vector giving the quantiles to calculate for each parameter (default = \code{c(0.025, 0.5, 0.975)})}

\item{ac_lags}{numeric vector of lags over which to calculate autocorrelation of samples within chains (default = \code{c(2,5,10)})}

\item{...}{further arguments passed to or from other methods}
}
\value{
a dataframe with summary statistics
}
\description{
This function takes a \code{nomad_model} object and calculates summary statistics
for the underlying \code{\link[mobility:mobility]{mobility::mobility()}}. This is a wrapper function of
\link[MCMCvis:MCMCsummary]{MCMCvis::MCMCsummary} that calculates summary statistics for each parameter
in a \code{\link[mobility:mobility]{mobility::mobility()}} object.
}
\details{
Summary statistics are calculated for all parameters across each chain along
with convergance diagnosics like the Gelman-Rubin convergence diagnostic and
(Rhat) and samples auto-correlation foreach parameter. If the model object
contains deviance and penalty parameters, then Deviance Information Criterion
(DIC) is calculated and appended to the summary.
}
\examples{
# Get nomad_model object
nmd_model <- nomad::model_db$zmb_cdr_2020_mod_dd_exp

# Check model fit
summary(nmd_model)
}
\seealso{
Other model: 
\code{\link[mobility]{check}()},
\code{\link[mobility]{compare}()},
\code{\link[mobility]{fit_jags}()},
\code{\link[mobility]{fit_prob_travel}()},
\code{\link[mobility]{mobility}()},
\code{\link[mobility]{predict}()},
\code{\link[mobility]{residuals}()}
}
\author{
John Giles
}
