% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops.R
\name{unpack_pop}
\alias{unpack_pop}
\title{Extract information from rasters}
\usage{
unpack_pop(iso3c_sf, pop)
}
\arguments{
\item{iso3c_sf}{A simple feature shape file to extract for}

\item{pop}{The pop raster from \code{\link{get_pop}}}
}
\value{
Tibble with list columns of raw extracted values
}
\description{
Extract information from rasters
}
