% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zambia_cdr_2020_mod_dd_exp}
\alias{zambia_cdr_2020_mod_dd_exp}
\title{Zambia 2020 Call Data Records Exponential Departure-Diffusion Model}
\format{
A \code{nomad::model} class object
}
\source{
\url{https://elifesciences.org/articles/68441}
}
\usage{
zambia_cdr_2020_mod_dd_exp
}
\description{
Mobility model generated using \code{\link[mobility:mobility]{mobility::mobility()}} based on call data
records collected in Zambia in 2020 The mobility model is an exponential
Departure-Diffusion model.
}
\details{
For full metadata of the CDR data, please see \code{?nomad::mobility_database}
}
\keyword{datasets}
