% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\alias{build_path}
\title{Build a nomad archive from github}
\usage{
build(ref, dest, progress = NULL)

build_path(path, dest, progress = NULL)
}
\arguments{
\item{ref}{A reference in \code{username/repo} format.  Branches
may be used.}

\item{dest}{Destination directory - must be an empty directory if
it exists, otherwise specify a new place.}

\item{progress}{Passed through to \code{nomad::pack}}

\item{path}{Path to source, when using \code{build_path}}
}
\description{
Build a nomad archive from a github repository
}
