% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{nomad_model_}
\alias{nomad_model_}
\title{R6 Class for mobility models}
\description{
A nomad mobility model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-nomad_model-new}{\code{nomad_model_$new()}}
\item \href{#method-nomad_model-predict}{\code{nomad_model_$predict()}}
\item \href{#method-nomad_model-check_coordinates}{\code{nomad_model_$check_coordinates()}}
\item \href{#method-nomad_model-remove_model_data}{\code{nomad_model_$remove_model_data()}}
\item \href{#method-nomad_model-get_model}{\code{nomad_model_$get_model()}}
\item \href{#method-nomad_model-get_data_name}{\code{nomad_model_$get_data_name()}}
\item \href{#method-nomad_model-get_data_ref}{\code{nomad_model_$get_data_ref()}}
\item \href{#method-nomad_model-get_data_fields}{\code{nomad_model_$get_data_fields()}}
\item \href{#method-nomad_model-get_check_res}{\code{nomad_model_$get_check_res()}}
\item \href{#method-nomad_model-set_check_res}{\code{nomad_model_$set_check_res()}}
\item \href{#method-nomad_model-set_model_data_ref}{\code{nomad_model_$set_model_data_ref()}}
\item \href{#method-nomad_model-get_model_name}{\code{nomad_model_$get_model_name()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{nomad} mobility model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$new(model, data_name, data_ref = NULL, data_fields = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{\code{\link[mobility:mobility]{mobility::mobility()}} model}

\item{\code{data_name}}{The model's name to link to \code{nomad::mobility_db}}

\item{\code{data_ref}}{Optional internal reference to shared package model data.}

\item{\code{data_fields}}{Optional character vector of shared data fields.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{nomad_model} object.
Prediction and simulation method for 'mobility.model' class

This function uses a fitted \code{mobility.model} object to simulate a
connectivity matrix based on estimated parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-predict"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$predict(newdata = NULL, nsim = 1, seed = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{a list containing new data to used in model prediction.
If \code{NULL} (the default) the function will simulate the model with
the data used for fitting. See \code{\link[mobility:predict]{mobility::predict()}}}

\item{\code{nsim}}{number of simulations (default = 1)}

\item{\code{seed}}{optional integer specifying the call to \code{set.seed}
prior to model simulation (default = \code{NULL})}

\item{\code{...}}{further arguments passed to or from other methods}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
When \code{nsim = 1}, the prediction matrix is calculated
using the mean point estimate of parameter values. If \code{nsim > 1}
then returns and array that contains \code{nsim} number of simulated
replications based on the posterior distributions of each parameter.
}

\subsection{Returns}{
a vector, matrix, or array containing predicted or
simulated mobility values.
Check coordinates

This function uses a fitted \code{mobility.model} object to simulate a
connectivity matrix based on estimated parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-check_coordinates"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-check_coordinates}{}}}
\subsection{Method \code{check_coordinates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$check_coordinates(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{List of \code{newdata} that is passed to
\link[mobility:predict]{mobility::predict}
Remove data model was trained on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-remove_model_data"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-remove_model_data}{}}}
\subsection{Method \code{remove_model_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$remove_model_data(
  M = FALSE,
  D = FALSE,
  N_orig = FALSE,
  N_dest = FALSE,
  S = FALSE,
  save_model_checks = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{Mobility Data}

\item{\code{D}}{Distance Data}

\item{\code{N_orig}}{Population Origin Data}

\item{\code{N_dest}}{Population Destination Data}

\item{\code{S}}{Radiation intervening opportunity matrix}

\item{\code{save_model_checks}}{Logical to save model check results. Default = FALSE}

\item{\code{...}}{further arguments passed to \code{\link[=png]{png()}} for saving model check plot
Get model object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$get_model(hydrate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hydrate}}{If \code{TRUE}, attach any shared package data referenced by
this model before returning it.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Underlying \code{\link[mobility:mobility]{mobility::mobility()}} model
Get data name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-get_data_name"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-get_data_name}{}}}
\subsection{Method \code{get_data_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$get_data_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data name
Get shared model data reference
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-get_data_ref"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-get_data_ref}{}}}
\subsection{Method \code{get_data_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$get_data_ref()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Shared data reference or \code{NULL}
Get shared model data fields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-get_data_fields"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-get_data_fields}{}}}
\subsection{Method \code{get_data_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$get_data_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of shared data fields or \code{NULL}
Get check results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-get_check_res"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-get_check_res}{}}}
\subsection{Method \code{get_check_res()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$get_check_res()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Results of model check
Set check results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-set_check_res"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-set_check_res}{}}}
\subsection{Method \code{set_check_res()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$set_check_res(check_res)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check_res}}{Results of model check
Set shared model data reference}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-set_model_data_ref"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-set_model_data_ref}{}}}
\subsection{Method \code{set_model_data_ref()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$set_model_data_ref(data_ref = NULL, data_fields = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_ref}}{Shared data reference}

\item{\code{data_fields}}{Shared model data fields
Get the model's name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nomad_model-get_model_name"></a>}}
\if{latex}{\out{\hypertarget{method-nomad_model-get_model_name}{}}}
\subsection{Method \code{get_model_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nomad_model_$get_model_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Model name
}
}
}
