% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{nomad_remove_model_data}
\alias{nomad_remove_model_data}
\title{Remove mobility data that mobility model was trained on}
\usage{
nomad_remove_model_data(
  model,
  M = FALSE,
  D = FALSE,
  N_orig = FALSE,
  N_dest = FALSE,
  save_model_checks = FALSE,
  ...
)
}
\arguments{
\item{model}{\code{\link[=nomad_model]{nomad_model()}} model}

\item{M}{Mobility Data. Default = FALSE (i.e. do not remove)}

\item{D}{Distance Data. Default = FALSE}

\item{N_orig}{Population Origin Data. Default = FALSE}

\item{N_dest}{Population Destination Data. Default = FALSE}

\item{save_model_checks}{Logical to save model check results. Default = FALSE}

\item{...}{further arguments passed to \code{\link[=png]{png()}} for saving model check plot}
}
\description{
Remove mobility data that mobility model was trained on
}
\details{
Specify \code{TRUE} to remove a data type.
}
