% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomad.R
\name{pack}
\alias{pack}
\title{Create recon usb stick}
\usage{
pack(path, progress = TRUE)
}
\arguments{
\item{path}{Path to download things to}

\item{progress}{Print a progress bar for each downloaded file}
}
\description{
Create recon usb stick
}
