% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.nomad_model}
\alias{predict.nomad_model}
\title{Prediction and simulation method for 'nomad_model' class}
\usage{
\method{predict}{nomad_model}(object, newdata = NULL, nsim = 1, seed = NULL, unit = NULL, ...)
}
\arguments{
\item{object}{a \code{\link[=nomad_model]{nomad_model()}} object containing the fitted mobility
model and its associated meta data}

\item{newdata}{a list containing new data to used in model prediction. If \code{NULL} (the default) the function will simulate the model with the data used for fitting.
\describe{
  \item{\code{D}}{a named matrix of distances between all \eqn{ij} location pairs}
  \item{\code{N}}{a named vector of population sizes for all locations (either N or both N_orig and N_dest must be supplied)}
  \item{\code{N_orig}}{a named vector of population sizes for each origin}
  \item{\code{N_dest}}{a named vector of population sizes for each destination}
}}

\item{nsim}{number of simulations (default = 1).}

\item{seed}{optional integer specifying the call to \code{set.seed} prior to model simulation (default = \code{NULL})}

\item{unit}{The unit of distance associated with the x and y coordinates.
Default = \code{NULL}, which assumes the provided coordinates are longitude (x)
and latitiude (y).}

\item{...}{further arguments passed to or from other methods}
}
\value{
a vector, matrix, or array containing predited or simulated mobility values.
}
\description{
This function uses the fitted \code{\link[mobility:mobility]{mobility::mobility()}} object inside the the
\code{\link[=nomad_model]{nomad_model()}} to simulate a connectivity matrix based on
estimated parameters.
}
\details{
When \code{nsim = 1}, the prediction matrix is calculated
using the mean point estimate of parameter values. If \code{nsim > 1}
then an array that contains \code{nsim} number of simulated replications
is returned based on the posterior distributions of each parameter.

TODO: Warnings are not currently implemented (Milestone 2)

Warnings will be shown if the \code{newdata} provided is unlikely to be well
predicted by the model. This will happen under two circumstances.
\enumerate{
\item If the coordinates provided are at a spatial scale that is unsuitable
given the spatial scale that the model was fit against. For example, if the
model was fit using data collecting mobility between locations that are
100s km apart on average but the coordinates are 1km apart on average.
\item If the coordinates provided span international borders. Currently, there
are no models available in \code{nomad} that were fit using mobility data that
includes international movements. As a result, any predictions made for
international locations may poorly capture international mobility.
}
}
\examples{
# Get nomad_model object
nmd_model <- nomad::model_db$zmb_cdr_2020_mod_dd_exp

# Produce model predictions
predict(nmd_model)

}
\seealso{
Other model: 
\code{\link[mobility]{check}()},
\code{\link[mobility]{compare}()},
\code{\link[mobility]{fit_jags}()},
\code{\link[mobility]{fit_prob_travel}()},
\code{\link[mobility]{mobility}()},
\code{\link[mobility]{residuals}()},
\code{\link[mobility]{summary}()}
}
\author{
John Giles
}
